change IdTruth from UShort_t to Int_t
Former-commit-id: 9c11d545dfe330f31c134cbc5aaa4c3dd5765303

+++ StRoot/RTS/src/DAQ_TPX/tpxFCF.h
@@ -73,7 +73,7 @@ struct tpxFCF_cl {
 	fcf_short p2 ;
 
 	fcf_short flags ;
-	unsigned short track_id ;
+	int       track_id ;
 
 	short quality ;	
 	short sim_length ;	// length of the corresponding sim data in *sim
